% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseArgs.R
\name{parseArgs}
\alias{parseArgs}
\title{Process arguments supplied to GLU by the user.}
\usage{
parseArgs()
}
\value{
List of options specified by the user.
}
\description{
Process arguments supplied to GLU by the user.
}
