% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGLUForDirectory.R
\name{runGLUForDirectory}
\alias{runGLUForDirectory}
\title{Run GLU for all files in a directory}
\usage{
runGLUForDirectory(indir, outdir = NULL, device = 0,
  daystart = "06:30", nightstart = "23:00",
  timeformat = "\%d/\%m/\%y \%H:\%M:\%S", impute = FALSE, freq = 5,
  outlierthreshold = 5, hypothreshold = NULL, hyperthreshold = NULL,
  save = FALSE, pregnancy = FALSE, diabetes = FALSE)
}
\arguments{
\item{indir}{Path of input directory containing input files.}

\item{outdir}{Path of output directory where derived data will be stored.}

\item{device}{Device type: 0 (medtronic ipro2); 1 (Dexcom G2); 2 (Abbott Freestyle Libre), default 0.}

\item{daystart}{Time of night to day transition in format HH:MM, default 06:30.}

\item{nightstart}{Time of day to night transition in format HH:MM, default 23:00.}

\item{timeformat}{Time format in the CGM data, default %d/%m/%y %H:%M:%S.}

\item{impute}{Logical. If TRUE then the 'approximal imputation' approach to dealing with missing data is used. If FALSE then the 'complete days' approach is used.}

\item{freq}{Integer. CGM epoch data frequency (minutes).}

\item{outlierthreshold}{Numeric. The value k used for outlier detection threshold d=k*SD.}

\item{hypothreshold}{Numeric. Threshold between hypo- and normo- glycaemia.}

\item{hyperthreshold}{Numeric. Threshold between normo- and hyper- glycaemia.}

\item{save}{Logical. If TRUE then derived CGM sequence(s) are stored.}

\item{pregnancy}{Logical. If TRUE then data is for pregnancy study, so pregnancy specific statistics should be derived.}

\item{diabetes}{Logical. If TRUE then data is for diabetes study, so pregnancy specific statistics should be derived.}
}
\description{
Run GLU for all files in a directory
}
